quiz/src/main/java/ua/nykyforov/service/quiz/application/service/UserServiceImpl.java
package ua.nykyforov.service.quiz.application.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ua.nykyforov.service.quiz.core.application.UserService;
import ua.nykyforov.service.quiz.core.dao.UserDAO;
import ua.nykyforov.service.quiz.core.model.User;

import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {

    private final UserDAO userDAO;

    @Autowired
    public UserServiceImpl(UserDAO userDAO) {
        this.userDAO = userDAO;
    }

    @Override
    public User save(User user) {
        return userDAO.save(user);
    }

    @Override
    public Optional<User> getById(long id) {
        return Optional.ofNullable(userDAO.getById(id));
    }
}
